+++ docs/database/add_contacts_email_constraint.sql
@@ -42,3 +42,4 @@ COMMENT ON INDEX contacts_franchise_email_unique_idx IS 'Unique constraint for u
 
 
 
+