+++ docs/database/add_county_to_territory_checks.sql
@@ -28,3 +28,4 @@ COMMENT ON COLUMN territory_checks.county IS 'County name for territory check lo
 
 
 
+