+++ docs/database/add_family_of_brands_tables.sql
@@ -84,3 +84,4 @@ COMMENT ON VIEW franchises_with_family IS 'Convenience view joining franchises w
 
 
 
+