+++ docs/database/add_llm_processed_at.sql
@@ -35,3 +35,4 @@ COMMENT ON COLUMN franchises.llm_processed_at IS 'Timestamp when franchise was p
 
 
 
+