+++ docs/database/add_territory_normalization_fields.sql
@@ -35,3 +35,4 @@ COMMENT ON COLUMN territory_checks.is_resale IS 'Whether this territory check is
 
 
 
+