+++ frontend/src/types/react-simple-maps.d.ts
@@ -70,3 +70,4 @@ declare module 'react-simple-maps' {
 
 
 
+